Proper formatting for order totals on transaction detail page

+++ customerapp/src/main/java/com/bokoup/customerapp/dom/model/BokoupTransaction.kt
@@ -2,7 +2,7 @@ package com.bokoup.customerapp.dom.model
 
 import com.dgsd.ksol.core.model.PublicKey
 import com.dgsd.ksol.core.model.TransactionSignature
-import java.text.DateFormatSymbols
+import java.math.BigDecimal
 import java.time.OffsetDateTime
 
 data class BokoupTransaction(
@@ -11,8 +11,8 @@ data class BokoupTransaction(
     val merchantName: String,
     val orderId: String?,
     val paymentId: String?,
-    val orderTotal: Int?,
-    val discountValue: Float?,
+    val orderTotal: BigDecimal?,
+    val discountValue: BigDecimal?,
     val type: Type,
     val tokenInfo: TokenInfo,
 ) {
